+++ app/src/test/java/com/codesoom/assignment/services/ProductServiceTest.java
@@ -3,7 +3,7 @@ package com.codesoom.assignment.services;
 import com.codesoom.assignment.ProductNotFoundException;
 import com.codesoom.assignment.services.product.domain.Product;
 import com.codesoom.assignment.services.product.domain.ProductRepository;
-import com.codesoom.assignment.dto.ProductData;
+import com.codesoom.assignment.services.product.dto.ProductData;
 import com.codesoom.assignment.services.product.ProductService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
